backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
transformers==4.37.2
torch==2.2.0
scikit-learn==1.4.0
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.1